requirements.txt
Pillow==2.4.0
numpy==1.8.1
scipy==0.14.0
